Refactor MeetingService.js to eliminate God Object and upgrade Gemini models
