+++ server/services/MeetingStorageService.js
@@ -1,0 +1,44 @@
+import Meeting from "../models/meetingModel.js";
+
+export const createMeetingRecord = async (data) => {
+  return await Meeting.create(data);
+};
+
+export const findMeetingById = async (id) => {
+  return await Meeting.findById(id);
+};
+
+export const findMeetingByQuery = async (query) => {
+  return await Meeting.findOne(query);
+};
+
+export const getMeetingsQuery = async (query, skip, limit) => {
+  return await Meeting.find(query)
+    .sort({ createdAt: -1 })
+    .skip(skip)
+    .limit(limit)
+    .select(
+      "title summary structuredMoM createdAt date meetingType status time duration recordingType organization",
+    )
+    .populate("organization", "name");
+};
+
+export const countMeetingsQuery = async (query) => {
+  return await Meeting.countDocuments(query);
+};
+
+export const deleteMeetingById = async (id) => {
+  return await Meeting.findByIdAndDelete(id);
+};
+
+export const searchMeetingsRecords = async (searchQuery) => {
+  return await Meeting.find({
+    $or: [
+      { title: { $regex: searchQuery, $options: "i" } },
+      { summary: { $regex: searchQuery, $options: "i" } },
+      { transcript: { $regex: searchQuery, $options: "i" } },
+    ],
+  })
+    .sort({ createdAt: -1 })
+    .select("title summary transcript createdAt date meetingType");
+};
