+++ server/services/PolicyService.js
@@ -31,7 +31,7 @@ const pdf = require("pdf-parse");
 
 // ── Config ─────────────────────────────────────────────────────
 const GEMINI_API_KEY = process.env.GEMINI_API_KEY;
-const GEMINI_MODEL = process.env.GEMINI_MODEL || "gemini-2.0-flash";
+const GEMINI_MODEL = process.env.GEMINI_MODEL || "gemini-2.5-flash";
 
 // ── Upload directory setup ─────────────────────────────────────
 const UPLOAD_DIR = path.resolve("uploads/policies");
