+++ server/services/policyComplianceService.js
@@ -26,7 +26,7 @@ import { embedText, initVectorStore } from "../utils/embeddingUtils.js";
 dotenv.config();
 
 const GEMINI_API_KEY = process.env.GEMINI_API_KEY;
-const GEMINI_MODEL = process.env.GEMINI_MODEL || "gemini-2.0-flash";
+const GEMINI_MODEL = process.env.GEMINI_MODEL || "gemini-2.5-flash";
 
 // Namespace policy vectors separately from meeting vectors so semantic
 // search over meetings never picks up policy text and vice versa.
