+++ server/utils/fileUtils.js
@@ -1,0 +1,11 @@
+import path from "path";
+
+export const validatePath = (filePath) => {
+  if (!filePath) throw new Error("Path is required");
+  const resolved = path.resolve(filePath);
+  const uploadsDir = path.resolve("uploads");
+  if (!resolved.startsWith(uploadsDir)) {
+    throw new Error("Directory traversal detected: Access denied");
+  }
+  return resolved;
+};
